remove legacy implementation
